Extract whitelisting functionality out into a reporter class
This reduces duplication in the checks for the price of
denying the checks direct access to the whitelist class.
Therefore we also extract the whitelist expiry check into a
non-checky special case report. Worth it, since whitelist
expiry checking is indeed a special case.
