+++ lib/check_runner.rb
@@ -10,7 +10,7 @@ class CheckRunner
     suppress_progress = env["SUPPRESS_PROGRESS"] ? true : false
 
     checker_db = CheckerDB.new(checker_db_name)
-    whitelist = Whitelist.load(whitelist_file)
+    @check_reporter = Checks::Reporter.new(Whitelist.load(whitelist_file))
 
     @progress_reporter = suppress_progress ? ProgressReporter.noop : ProgressReporter.new
 
@@ -20,22 +20,23 @@ class CheckRunner
       @importers << Import::PublishingApiImporter.new(checker_db, @progress_reporter, publishing_api_url)
     end
 
-    @checks = load_checks(checker_db, whitelist, check_names)
+    @checks = load_checks(checker_db, check_names)
   end
 
   def run
     run_importers
-    reports = run_checks
-    report_results(reports)
+    check_reports = run_checks
+    whitelist_expiry_report = @check_reporter.report_expired_whitelist_entries
+    report_results(check_reports + [whitelist_expiry_report])
   end
 
 private
 
-  def load_checks(checker_db, whitelist, check_names)
+  def load_checks(checker_db, check_names)
     check_files = Dir[File.join(File.dirname(__FILE__), 'checks', '*.rb')]
     check_names_to_run = check_names.empty? ? get_check_names(check_files) : check_names
     check_files.each { |file| require file }
-    check_names_to_run.map { |check_name| Checks.const_get(check_name).new(check_name, checker_db, whitelist) }
+    check_names_to_run.map { |check_name| Checks.const_get(check_name).new(check_name, checker_db, @check_reporter) }
   end
 
   def get_check_names(check_files)
