+++ lib/checks/base_paths_missing_from_publishing_api.rb
@@ -1,9 +1,9 @@
 module Checks
   class BasePathsMissingFromPublishingApi
-    def initialize(name, checker_db, whitelist)
+    def initialize(name, checker_db, reporter)
       @name = name
       @checker_db = checker_db
-      @whitelist = whitelist
+      @reporter = reporter
     end
 
     def run_check
@@ -21,9 +21,7 @@ module Checks
 
       headers = %w(base_path format index document_type)
       rows = @checker_db.execute(query)
-      whitelist_function = @whitelist.get_whitelist_function(@name, headers)
-
-      Report.create(@name, headers, rows, whitelist_function)
+      @reporter.create_report(@name, headers, rows)
     end
   end
 end
